app/requirements.txt
absl-py
ale-py
astunparse
blinker
certifi
charset-normalizer
click
cloudpickle
contourpy
cycler
dill
Farama-Notifications
filelock
Flask
flask-cors
flatbuffers
fonttools
fsspec
gast
google-pasta
grpcio
gymnasium
h5py
idna
itsdangerous
Jinja2
joblib
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
mpmath
multiprocess
namex
networkx
numpy
opencv-python
opt_einsum
optree
packaging
pandas
pathos
pillow
pox
ppft
protobuf
psutil
psycopg2-binary
pygame
pyglet
Pygments
pyparsing
python-dateutil
pytz
requests
rich
scikit-learn
scipy
setuptools
simglucose
six
stable_baselines3
sympy
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
torch
tqdm
typing_extensions
tzdata
urllib3
Werkzeug
wheel
wrapt
